Refactoring 6 Files
